Absorb render into presentation facade
