+++ crates/perfgate-app/Cargo.toml
@@ -17,10 +17,11 @@ categories = ["development-tools"]
 perfgate-types.workspace = true
 perfgate-domain.workspace = true
 perfgate-adapters.workspace = true
-perfgate-render.workspace = true
 perfgate-export.workspace = true
 perfgate-sensor.workspace = true
 anyhow.workspace = true
+glob.workspace = true
+handlebars = "6.3"
 humantime.workspace = true
 serde.workspace = true
 serde_json.workspace = true
