+++ crates/perfgate-github/Cargo.toml
@@ -13,8 +13,8 @@ categories = ["development-tools"]
 
 [dependencies]
 anyhow.workspace = true
+perfgate-app.workspace = true
 perfgate-types.workspace = true
-perfgate-render.workspace = true
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls-webpki-roots"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
