+++ crates/perfgate-render/Cargo.toml
@@ -7,23 +7,19 @@ license.workspace = true
 repository.workspace = true
 homepage.workspace = true
 authors.workspace = true
-description = "Rendering utilities for perfgate output (markdown, GitHub annotations)"
+description = "Workspace-only compatibility wrapper for perfgate::presentation::render"
 keywords = ["markdown", "github", "annotations", "rendering"]
 categories = ["development-tools"]
 documentation = "https://docs.rs/perfgate-render"
 readme = "README.md"
+publish = false
 
 [dependencies]
-perfgate-types.workspace = true
-serde_json.workspace = true
-anyhow.workspace = true
-handlebars = "6.3"
-glob.workspace = true
-tempfile = "3.26.0"
+perfgate-app.workspace = true
 
 [dev-dependencies]
-proptest = "1.7"
-insta.workspace = true
+perfgate.workspace = true
+perfgate-types.workspace = true
 
 [lints]
 workspace = true
