+++ docs/CRATE_SEAMS.md
@@ -40,10 +40,11 @@ PR #223 started the real collapse and is the current implementation truth:
 |--------------|---------------|--------|
 | `perfgate-validation` | `perfgate_types::validation` | crate deleted |
 | `perfgate-auth` | `perfgate_api::auth` | crate deleted |
-| `perfgate-summary` | `perfgate_render::summary` | crate deleted |
+| `perfgate-summary` | `perfgate::presentation::render::summary` | crate deleted |
 | `perfgate-stats` | `perfgate_domain::stats` | crate deleted |
 | `perfgate-paired` | `perfgate_domain::paired` | workspace-only compatibility wrapper, `publish = false` |
 | `perfgate-error` | `perfgate_types::error` | workspace-only compatibility wrapper, `publish = false` |
+| `perfgate-render` | `perfgate::presentation::render` | workspace-only compatibility wrapper, `publish = false` |
 | `perfgate-fake` | private workspace crate | marked `publish = false` |
 | `perfgate-profile` | `perfgate::runtime::profile` | crate deleted |
 | `perfgate-ingest` | `perfgate::integrations::ingest` | crate deleted |
@@ -67,7 +68,6 @@ high-level target is:
 | `perfgate-config` | `perfgate_types::config` and `perfgate_client::ResolvedServerConfig` |
 | `perfgate-api` | `perfgate_types::baseline_service` or shared client/server contract |
 | `perfgate-domain` | `perfgate::domain` |
-| `perfgate-render` | `perfgate::presentation::render` |
 | `perfgate-export` | `perfgate::presentation::export` |
 | `perfgate-sensor` | `perfgate::presentation::sensor` |
 | `perfgate-adapters` | `perfgate::runtime` |
