+++ docs/WORKSPACE.md
@@ -8,7 +8,7 @@ This file is automatically generated by `cargo run -p xtask -- docs-sync`.
 |-------|-------------|------------------|
 | `perfgate-error` | Compatibility wrapper for perfgate_types::error | 100% |
 | `perfgate-export` | Export formats (CSV, JSONL, HTML, Prometheus) | 90% |
-| `perfgate-render` | Markdown and GitHub annotations rendering | 90% |
+| `perfgate-render` | Workspace-only compatibility wrapper for perfgate::presentation::render | 90% |
 | `perfgate-sensor` | Sensor report builder for cockpit integration | 90% |
 | `perfgate-paired` | Paired benchmarking statistics (A/B testing) | 100% |
 | `perfgate-fake` | Test utilities and fake implementations | 70% |
